Move OmniLogicConfigEntry to coordinator module
